Move weather fetcher into GeoShot.Web service
