+++ GeoShot.Web/Features/Weather/WeatherData.cs
@@ -1,0 +1,5 @@
+using System.Text.Json;
+
+namespace GeoShot.Web.Features.Weather;
+
+public record WeatherData(double Latitude, double Longitude, JsonElement Data);
